requirements.txt
transformers
datasets
scikit-learn
pandas
numpy
pytorch-lightning
wandb
matplotlib
tqdm

# pip install -r requirements.txt